Avoid deletion of yum cache during concurrent executions
By placing the cleanup of the mock caches and the chroot initialization
in a critical region, we make sure the yum cache is not deleted while
another build is initializing it's chroot. After the initialization, the
chroot is ready in a unique path and there's no need to access the
caches anymore.

This allows build-iso to be executed in parallel.