Add missing minimal packages
Pungi requires some packages to be available when preparing the ISO root file
system, so it can download them from the configured repositories. Since we
intend to create a local minimal repository instead of passing the CentOS ones
(to enable a higher priority for Host OS packages), they must be mentioned in
our configuration to be downloaded to the chroot.

The anaconda package has authconfig and chrony as dependencies, so those
packages do not need to be listed explicitly.